// src/database/config/database-config.type.ts
export type DatabaseConfig = {
  url?: string;
  type: string;
  host?: string;
  port?: number;
  password?: string;
  name?: string;
  username?: string;
  synchronize: boolean;
  maxConnections: number;
  sslEnabled: boolean;
  rejectUnauthorized: boolean;
  ca?: string;
  key?: string;
  cert?: string;
};